services/transcriber/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
Pillow>=10.0.0
